revert reduced asterix, dep fix
